+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/fir/generators/ContributedInterfaceSupertypeGenerator.kt
@@ -3,7 +3,6 @@
 package dev.zacsweers.metro.compiler.fir.generators
 
 import dev.zacsweers.metro.compiler.Symbols
-import dev.zacsweers.metro.compiler.expectAsOrNull
 import dev.zacsweers.metro.compiler.fir.FirTypeKey
 import dev.zacsweers.metro.compiler.fir.annotationsIn
 import dev.zacsweers.metro.compiler.fir.anvilKClassBoundTypeArgument
@@ -17,7 +16,6 @@ import dev.zacsweers.metro.compiler.fir.resolvedExcludedClassIds
 import dev.zacsweers.metro.compiler.fir.resolvedReplacedClassIds
 import dev.zacsweers.metro.compiler.fir.resolvedScopeClassId
 import dev.zacsweers.metro.compiler.fir.scopeArgument
-import dev.zacsweers.metro.compiler.flatMapToSet
 import dev.zacsweers.metro.compiler.singleOrError
 import java.util.TreeMap
 import org.jetbrains.kotlin.fir.FirAnnotationContainer
@@ -26,11 +24,9 @@ import org.jetbrains.kotlin.fir.analysis.checkers.getContainingClassSymbol
 import org.jetbrains.kotlin.fir.caches.FirCache
 import org.jetbrains.kotlin.fir.caches.firCachesFactory
 import org.jetbrains.kotlin.fir.declarations.FirClassLikeDeclaration
-import org.jetbrains.kotlin.fir.declarations.FirDeclarationOrigin
 import org.jetbrains.kotlin.fir.declarations.FirResolvePhase
 import org.jetbrains.kotlin.fir.declarations.ResolveStateAccess
 import org.jetbrains.kotlin.fir.expressions.FirAnnotation
-import org.jetbrains.kotlin.fir.expressions.FirArrayLiteral
 import org.jetbrains.kotlin.fir.extensions.FirDeclarationPredicateRegistrar
 import org.jetbrains.kotlin.fir.extensions.FirSupertypeGenerationExtension
 import org.jetbrains.kotlin.fir.extensions.predicateBasedProvider
@@ -40,6 +36,8 @@ import org.jetbrains.kotlin.fir.recordFqNameLookup
 import org.jetbrains.kotlin.fir.resolve.providers.symbolProvider
 import org.jetbrains.kotlin.fir.resolve.toClassSymbol
 import org.jetbrains.kotlin.fir.resolve.toRegularClassSymbol
+import org.jetbrains.kotlin.fir.scopes.getSingleClassifier
+import org.jetbrains.kotlin.fir.scopes.impl.declaredMemberScope
 import org.jetbrains.kotlin.fir.symbols.SymbolInternals
 import org.jetbrains.kotlin.fir.symbols.impl.FirClassLikeSymbol
 import org.jetbrains.kotlin.fir.symbols.impl.FirClassSymbol
@@ -75,20 +73,20 @@ internal class ContributedInterfaceSupertypeGenerator(session: FirSession) :
       val allSessions =
         sequenceOf(session).plus(session.moduleData.allDependsOnDependencies.map { it.session })
 
-      allSessions
-        .flatMap {
-          it.predicateBasedProvider.getSymbolsByPredicate(
-            session.predicates.contributesAnnotationPredicate
-          )
-        }
-        .filterIsInstance<FirRegularClassSymbol>()
-        .flatMapToSet { clazz ->
-          clazz.annotations
-            .annotationsIn(session, session.classIds.allContributesAnnotations)
-            .mapNotNull { it.resolvedScopeClassId(typeResolver) }
-            .filter { it == scopeClassId }
-            .map { clazz.classId.createNestedClassId(Symbols.Names.metroContribution) }
-        }
+      // Predicates can't see the generated $$MetroContribution classes, but we can access them
+      // by first querying the top level @ContributeX-annotated source symbols and then checking
+      // their declaration scopes
+      val contributingClasses =
+        allSessions
+          .flatMap {
+            it.predicateBasedProvider.getSymbolsByPredicate(
+              session.predicates.contributesAnnotationPredicate
+            )
+          }
+          .filterIsInstance<FirRegularClassSymbol>()
+          .toList()
+
+      getScopedContributions(contributingClasses, typeResolver)[scopeClassId].orEmpty()
     }
 
   private val generatedScopesToContributions:
@@ -101,50 +99,56 @@ internal class ContributedInterfaceSupertypeGenerator(session: FirSession) :
           scopeHintFqName.shortName(),
         )
 
-      buildMap<ClassId, MutableSet<ClassId>> {
-        for (contribution in functionsInPackage) {
-          // Origin is the single value param
-          val originClass =
-            contribution.valueParameterSymbols
-              .single()
-              .resolvedReturnType
-              .toRegularClassSymbol(session) ?: continue
-
-          originClass
-            .contributesAnnotations()
-            .mapNotNull { it.resolvedScopeClassId(typeResolver) }
-            .distinct()
-            .forEach { scopeClassId ->
-              val metroContribution =
-                originClass.classId.createNestedClassId(Symbols.Names.metroContribution)
-              getOrPut(scopeClassId, ::mutableSetOf).add(metroContribution)
-            }
+      val contributingClasses =
+        functionsInPackage.mapNotNull { contribution ->
+          // This is the single value param
+          contribution.valueParameterSymbols
+            .single()
+            .resolvedReturnType
+            .toRegularClassSymbol(session)
         }
-      }
+
+      getScopedContributions(contributingClasses, typeResolver)
     }
 
-  private fun FirRegularClassSymbol.contributesAnnotations(): Sequence<FirAnnotation> {
-    return annotations.annotationsIn(session, session.classIds.allContributesAnnotations).ifEmpty {
-      if (origin == FirDeclarationOrigin.Library) {
-        // When there are multiple contribution annotations, they end up getting bundled into a
-        // 'Container' type that we need to unwrap. This behavior appears to only occur for
-        // contributions from other modules/compilations.
-        annotations
-          .firstOrNull()
-          ?.argumentMapping
-          ?.mapping
-          ?.values
-          ?.firstOrNull()
-          ?.expectAsOrNull<FirArrayLiteral>()
-          ?.argumentList
-          ?.arguments
-          ?.expectAsOrNull<List<FirAnnotation>>()
-          .orEmpty()
-          .annotationsIn(session, session.classIds.allContributesAnnotations)
-      } else {
-        emptySequence<FirAnnotation>()
-      }
+  /**
+   * @param contributingClasses The classes annotated with some number of @ContributesX annotations.
+   * @return A mapping of scope ids to @MetroContribution-annotated nested classes.
+   */
+  private fun getScopedContributions(
+    contributingClasses: List<FirRegularClassSymbol>,
+    typeResolver: TypeResolveService,
+  ): Map<ClassId, Set<ClassId>> {
+    val scopesToNestedContributions = mutableMapOf<ClassId, MutableSet<ClassId>>()
+
+    contributingClasses.forEach { originClass ->
+      val classDeclarationContainer =
+        originClass.fir.symbol.declaredMemberScope(session, memberRequiredPhase = null)
+
+      val contributionNames =
+        classDeclarationContainer.getClassifierNames().filter {
+          it.identifier.startsWith(Symbols.Names.metroContribution.identifier)
+        }
+
+      contributionNames
+        .mapNotNull { nestedClassName ->
+          val nestedClass = classDeclarationContainer.getSingleClassifier(nestedClassName)
+
+          nestedClass
+            ?.annotations
+            ?.annotationsIn(session, setOf(Symbols.ClassIds.metroContribution))
+            ?.single()
+            ?.resolvedScopeClassId(typeResolver)
+            ?.let { scopeId -> scopeId to originClass.classId.createNestedClassId(nestedClassName) }
+        }
+        .forEach { (scopeClassId, nestedContributionId) ->
+          scopesToNestedContributions
+            .getOrPut(scopeClassId, ::mutableSetOf)
+            .add(nestedContributionId)
+        }
     }
+
+    return scopesToNestedContributions
   }
 
   private fun FirAnnotationContainer.graphAnnotation(): FirAnnotation? {