+++ compiler/src/test/kotlin/dev/zacsweers/metro/compiler/fir/AggregationTest.kt
@@ -940,19 +940,41 @@ class AggregationTest : MetroCompilerTest() {
 
       val altGraphClass = classLoader.loadClass("test.AltGraph")
       val altGraph = altGraphClass.generatedMetroGraphClass().createGraphWithNoArgs()
-      altGraphClass.assertHasContributedSupertype("test.ContributedInterface")
+      altGraphClass.assertHasContributedSupertype(
+        "test.ContributedInterface",
+        contributionNumber = 2,
+      )
       assertThat(altGraph.callProperty<String>("altVal")).isEqualTo("Hello, world!")
 
       val thirdGraphClass = classLoader.loadClass("test.ThirdGraph")
       val thirdGraph = thirdGraphClass.generatedMetroGraphClass().createGraphWithNoArgs()
-      thirdGraphClass.assertHasContributedSupertype("test.ContributedInterface")
+      thirdGraphClass.assertHasContributedSupertype(
+        "test.ContributedInterface",
+        contributionNumber = 3,
+      )
       assertThat(thirdGraph.callProperty<String>("thirdVal")).isEqualTo("Hello, world!")
     }
   }
 
-  private fun Class<*>.assertHasContributedSupertype(superTypeFqName: String) {
+  /**
+   * @param contributionNumber Represents which nested class is expected. Each nested contribution
+   *   class is suffixed with a number when it's created depending on how many scopes are
+   *   contributed to. E.g.
+   *
+   * ```
+   * @ContributesBinding(AppScope::class) // This maps to $$MetroContribution (technically number 1)
+   * @ContributesBinding(AltScope::class) // This maps to $$MetroContribution2
+   * @Inject
+   * class ContributingClass : SomeInterface
+   * ```
+   */
+  private fun Class<*>.assertHasContributedSupertype(
+    superTypeFqName: String,
+    contributionNumber: Int = 1,
+  ) {
+    val contributionSuffix = if (contributionNumber == 1) "" else contributionNumber.toString()
     assertThat(allSupertypes().map { it.name })
-      .containsExactly("$superTypeFqName$$\$MetroContribution", superTypeFqName)
+      .containsExactly("$superTypeFqName$$\$MetroContribution$contributionSuffix", superTypeFqName)
   }
 
   @Test
@@ -989,7 +1011,10 @@ class AggregationTest : MetroCompilerTest() {
 
       val altGraphClass = classLoader.loadClass("test.AltGraph")
       val altGraph = altGraphClass.generatedMetroGraphClass().createGraphWithNoArgs()
-      altGraphClass.assertHasContributedSupertype("test.ContributedInterface")
+      altGraphClass.assertHasContributedSupertype(
+        "test.ContributedInterface",
+        contributionNumber = 2,
+      )
       assertThat(altGraph.callProperty<String>("altVal")).isEqualTo("Hello, world!")
     }
   }
@@ -1530,6 +1555,272 @@ class AggregationTest : MetroCompilerTest() {
     ) {
       assertDiagnostics(
         "e: Impl.kt:7:46 Redundant explicit bound type test.Impl is the same as the annotated class test.Impl."
+      )
+    }
+  }
+
+  @Test
+  fun `repeated ContributesBinding supports multiple scopes for a single type in a merging module`() {
+    compile(
+      source(
+        """
+          abstract class AltScope private constructor()
+
+          interface ContributedInterface
+
+          @ContributesBinding(AppScope::class)
+          @ContributesBinding(AltScope::class)
+          @Inject
+          class Impl : ContributedInterface
+
+          @DependencyGraph(scope = AppScope::class)
+          interface ExampleGraph {
+            val contributedInterface: ContributedInterface
+          }
+
+          @DependencyGraph(scope = AltScope::class)
+          interface AltGraph {
+            val contributedInterface: ContributedInterface
+          }
+        """
+          .trimIndent()
+      )
+    ) {
+      val graph = ExampleGraph.generatedMetroGraphClass().createGraphWithNoArgs()
+      val contributedInterface = graph.callProperty<Any>("contributedInterface")
+      assertThat(contributedInterface).isNotNull()
+      assertThat(contributedInterface.javaClass.name).isEqualTo("test.Impl")
+
+      val altGraph =
+        classLoader.loadClass("test.AltGraph").generatedMetroGraphClass().createGraphWithNoArgs()
+      val altContributedInterface = altGraph.callProperty<Any>("contributedInterface")
+      assertThat(altContributedInterface).isNotNull()
+      assertThat(altContributedInterface.javaClass.name).isEqualTo("test.Impl")
+    }
+  }
+
+  @Test
+  fun `repeated ContributesBinding supports multiple scopes for a single type in a downstream module`() {
+    val previousCompilation =
+      compile(
+        source(
+          """
+          abstract class AltScope private constructor()
+
+          interface ContributedInterface
+
+          @ContributesBinding(AppScope::class)
+          @ContributesBinding(AltScope::class)
+          @Inject
+          class Impl : ContributedInterface
+        """
+            .trimIndent()
+        )
+      )
+
+    compile(
+      source(
+        """
+          @DependencyGraph(scope = AppScope::class)
+          interface ExampleGraph {
+            val contributedInterface: ContributedInterface
+          }
+
+          @DependencyGraph(scope = AltScope::class)
+          interface AltGraph {
+            val contributedInterface: ContributedInterface
+          }
+        """
+          .trimIndent()
+      ),
+      previousCompilationResult = previousCompilation,
+    ) {
+      val graph = ExampleGraph.generatedMetroGraphClass().createGraphWithNoArgs()
+      val contributedInterface = graph.callProperty<Any>("contributedInterface")
+      assertThat(contributedInterface).isNotNull()
+      assertThat(contributedInterface.javaClass.name).isEqualTo("test.Impl")
+
+      val altGraph =
+        classLoader.loadClass("test.AltGraph").generatedMetroGraphClass().createGraphWithNoArgs()
+      val altContributedInterface = altGraph.callProperty<Any>("contributedInterface")
+      assertThat(altContributedInterface).isNotNull()
+      assertThat(altContributedInterface.javaClass.name).isEqualTo("test.Impl")
+    }
+  }
+
+  @Test
+  fun `repeated ContributesBinding supports multiple scopes for a different type in a merging module`() {
+    compile(
+      source(
+        """
+          abstract class AltScope private constructor()
+
+          interface ContributedInterface
+          interface OtherInterface
+
+          @ContributesBinding(AppScope::class, binding = binding<ContributedInterface>())
+          @ContributesBinding(AltScope::class, binding = binding<OtherInterface>())
+          @Inject
+          class Impl : ContributedInterface, OtherInterface
+
+          @DependencyGraph(scope = AppScope::class)
+          interface ExampleGraph {
+            val contributedInterface: ContributedInterface
+          }
+
+          @DependencyGraph(scope = AltScope::class)
+          interface AltGraph {
+            val otherInterface: OtherInterface
+          }
+        """
+          .trimIndent()
+      )
+    ) {
+      val graph = ExampleGraph.generatedMetroGraphClass().createGraphWithNoArgs()
+      val contributedInterface = graph.callProperty<Any>("contributedInterface")
+      assertThat(contributedInterface).isNotNull()
+      assertThat(contributedInterface.javaClass.name).isEqualTo("test.Impl")
+
+      val altGraph =
+        classLoader.loadClass("test.AltGraph").generatedMetroGraphClass().createGraphWithNoArgs()
+      val altContributedInterface = altGraph.callProperty<Any>("otherInterface")
+      assertThat(altContributedInterface).isNotNull()
+      assertThat(altContributedInterface.javaClass.name).isEqualTo("test.Impl")
+    }
+  }
+
+  @Test
+  fun `repeated ContributesBinding supports multiple scopes for a different type in a downstream module`() {
+    val previousCompilation =
+      compile(
+        source(
+          """
+          abstract class AltScope private constructor()
+
+          interface ContributedInterface
+          interface OtherInterface
+
+          @ContributesBinding(AppScope::class, binding = binding<ContributedInterface>())
+          @ContributesBinding(AltScope::class, binding = binding<OtherInterface>())
+          @Inject
+          class Impl : ContributedInterface, OtherInterface
+        """
+            .trimIndent()
+        )
+      )
+
+    compile(
+      source(
+        """
+          @DependencyGraph(scope = AppScope::class)
+          interface ExampleGraph {
+            val contributedInterface: ContributedInterface
+          }
+
+          @DependencyGraph(scope = AltScope::class)
+          interface AltGraph {
+            val otherInterface: OtherInterface
+          }
+        """
+          .trimIndent()
+      ),
+      previousCompilationResult = previousCompilation,
+    ) {
+      val graph = ExampleGraph.generatedMetroGraphClass().createGraphWithNoArgs()
+      val contributedInterface = graph.callProperty<Any>("contributedInterface")
+      assertThat(contributedInterface).isNotNull()
+      assertThat(contributedInterface.javaClass.name).isEqualTo("test.Impl")
+
+      val altGraph =
+        classLoader.loadClass("test.AltGraph").generatedMetroGraphClass().createGraphWithNoArgs()
+      val altContributedInterface = altGraph.callProperty<Any>("otherInterface")
+      assertThat(altContributedInterface).isNotNull()
+      assertThat(altContributedInterface.javaClass.name).isEqualTo("test.Impl")
+    }
+  }
+
+  @Test
+  fun `repeated ContributesBinding supports multiple scopes with a qualifier difference in a merging module`() {
+    compile(
+      source(
+        """
+          abstract class AltScope private constructor()
+
+          interface ContributedInterface
+
+          @ContributesBinding(AppScope::class)
+          @ContributesBinding(AltScope::class, binding = binding<@Named("Alt") ContributedInterface>())
+          @Inject
+          class Impl : ContributedInterface
+
+          @DependencyGraph(scope = AppScope::class)
+          interface ExampleGraph {
+            val contributedInterface: ContributedInterface
+          }
+
+          @DependencyGraph(scope = AltScope::class)
+          interface AltGraph {
+            @Named("Alt")
+            val otherInterface: ContributedInterface
+          }
+        """
+          .trimIndent()
+      )
+    ) {
+      val graph = ExampleGraph.generatedMetroGraphClass().createGraphWithNoArgs()
+      val contributedInterface = graph.callProperty<Any>("contributedInterface")
+      assertThat(contributedInterface).isNotNull()
+      assertThat(contributedInterface.javaClass.name).isEqualTo("test.Impl")
+
+      val altGraph =
+        classLoader.loadClass("test.AltGraph").generatedMetroGraphClass().createGraphWithNoArgs()
+      val altContributedInterface = altGraph.callProperty<Any>("otherInterface")
+      assertThat(altContributedInterface).isNotNull()
+      assertThat(altContributedInterface.javaClass.name).isEqualTo("test.Impl")
+    }
+  }
+
+  @Test
+  fun `repeated ContributesBinding supports multiple scopes for a different type without leaking the bindings`() {
+    compile(
+      source(
+        """
+          abstract class AltScope private constructor()
+
+          interface ContributedInterface
+          interface OtherInterface
+
+          @ContributesBinding(AppScope::class, binding = binding<ContributedInterface>())
+          @ContributesBinding(AltScope::class, binding = binding<OtherInterface>())
+          @Inject
+          class Impl : ContributedInterface, OtherInterface
+
+          @DependencyGraph(scope = AppScope::class)
+          interface ExampleGraph {
+            val contributedInterface: ContributedInterface
+          }
+
+          @DependencyGraph(scope = AltScope::class)
+          interface AltGraph {
+            val contributedInterface: ContributedInterface
+            val otherInterface: OtherInterface
+          }
+        """
+          .trimIndent()
+      ),
+      expectedExitCode = KotlinCompilation.ExitCode.COMPILATION_ERROR,
+    ) {
+      assertDiagnostics(
+        """
+          e: AltScope.kt:24:3 [Metro/MissingBinding] Cannot find an @Inject constructor or @Provides-annotated function/property for: test.ContributedInterface
+
+        test.ContributedInterface is requested at
+            [test.AltGraph] test.AltGraph#contributedInterface
+
+    Similar bindings:
+      - Impl (Subtype). Type: ConstructorInjected. Source: AltScope.kt:12:1
+        """
+          .trimIndent()
       )
     }
   }
@@ -2716,11 +3007,7 @@ class AggregationTest : MetroCompilerTest() {
       )
     ) {
       val graph = ExampleGraph
-      assertThat(graph.allSupertypes().map { it.name })
-        .containsExactly(
-          "test.ContributedInterface2",
-          "test.ContributedInterface2$$\$MetroContribution",
-        )
+      assertThat(graph.assertHasContributedSupertype("test.ContributedInterface2"))
     }
   }